refactor: Move aquery and cquery to be there own verbs under pkg/aspect
