+++ cmd/aspect/aquery/BUILD.bazel
@@ -6,7 +6,7 @@ go_library(
     importpath = "aspect.build/cli/cmd/aspect/aquery",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/aspect/query/aquery",
+        "//pkg/aspect/aquery",
         "//pkg/bazel",
         "//pkg/ioutils",
         "@com_github_spf13_cobra//:cobra",
