+++ pkg/aspect/cquery/cquery.go
@@ -4,17 +4,17 @@ Copyright © 2021 Aspect Build Systems Inc
 Not licensed for re-use.
 */
 
-package aquery
+package cquery
 
 import (
 	"github.com/spf13/cobra"
 
-	shared "aspect.build/cli/pkg/aspect/query"
+	"aspect.build/cli/pkg/aspect/query/shared"
 	"aspect.build/cli/pkg/bazel"
 	"aspect.build/cli/pkg/ioutils"
 )
 
-type AQuery struct {
+type CQuery struct {
 	ioutils.Streams
 
 	Bzl           bazel.Bazel
@@ -27,10 +27,10 @@ type AQuery struct {
 	Select       func(presetNames []string) shared.SelectRunner
 }
 
-func New(streams ioutils.Streams, bzl bazel.Bazel, isInteractive bool) *AQuery {
-	presets := shared.GetPrecannedQueries("aquery")
+func New(streams ioutils.Streams, bzl bazel.Bazel, isInteractive bool) *CQuery {
+	presets := shared.GetPrecannedQueries("cquery")
 
-	return &AQuery{
+	return &CQuery{
 		Streams:       streams,
 		Bzl:           bzl,
 		IsInteractive: isInteractive,
@@ -41,8 +41,8 @@ func New(streams ioutils.Streams, bzl bazel.Bazel, isInteractive bool) *AQuery {
 	}
 }
 
-func (q *AQuery) Run(cmd *cobra.Command, args []string) error {
-	verb := "aquery"
+func (q *CQuery) Run(cmd *cobra.Command, args []string) error {
+	verb := "cquery"
 	presets, presetNames, err := shared.ProcessQueries(q.Presets)
 	if err != nil {
 		return shared.GetPrettyError(cmd, err)
