--- pkg/aspect/query/query/BUILD.bazel
@@ -1,31 +1,0 @@
-load("@io_bazel_rules_go//go:def.bzl", "go_library", "go_test")
-
-go_library(
-    name = "query",
-    srcs = ["query.go"],
-    importpath = "aspect.build/cli/pkg/aspect/query/query",
-    visibility = ["//visibility:public"],
-    deps = [
-        "//pkg/aspect/query",
-        "//pkg/bazel",
-        "//pkg/ioutils",
-        "@com_github_spf13_cobra//:cobra",
-        "@com_github_spf13_viper//:viper",
-    ],
-)
-
-go_test(
-    name = "query_test",
-    srcs = ["query_test.go"],
-    deps = [
-        ":query",
-        "//pkg/aspect/query",
-        "//pkg/aspect/query/mock",
-        "//pkg/bazel/mock",
-        "//pkg/ioutils",
-        "@com_github_golang_mock//gomock",
-        "@com_github_onsi_gomega//:gomega",
-        "@com_github_spf13_cobra//:cobra",
-        "@com_github_spf13_viper//:viper",
-    ],
-)
