+++ pkg/aspect/query/query_test.go
@@ -4,20 +4,22 @@ Copyright © 2021 Aspect Build Systems Inc
 Not licensed for re-use.
 */
 
-package cquery_test
+package query_test
 
 import (
 	"fmt"
+	"os"
 	"strings"
 	"testing"
 
 	"github.com/golang/mock/gomock"
 	. "github.com/onsi/gomega"
 	"github.com/spf13/cobra"
+	"github.com/spf13/viper"
 
-	shared "aspect.build/cli/pkg/aspect/query"
-	"aspect.build/cli/pkg/aspect/query/cquery"
-	query_mock "aspect.build/cli/pkg/aspect/query/mock"
+	"aspect.build/cli/pkg/aspect/query"
+	"aspect.build/cli/pkg/aspect/query/shared"
+	query_mock "aspect.build/cli/pkg/aspect/query/shared/mock"
 	bazel_mock "aspect.build/cli/pkg/bazel/mock"
 	"aspect.build/cli/pkg/ioutils"
 )
@@ -31,20 +33,27 @@ func TestQuery(t *testing.T) {
 		spawner := bazel_mock.NewMockBazel(ctrl)
 		spawner.
 			EXPECT().
-			Spawn([]string{"cquery", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
+			Spawn([]string{"query", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
 			Return(0, nil)
 
 		var stdout strings.Builder
 		streams := ioutils.Streams{Stdout: &stdout}
-		q := cquery.New(streams, spawner, true)
+		q := query.New(streams, spawner, true)
 		q.Presets = []*shared.PresetQuery{
 			{
 				Name:        "why",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?target, ?dependency)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 		}
+
+		viper := *viper.New()
+		cfg, err := os.CreateTemp(os.Getenv("TEST_TMPDIR"), "cfg***.ini")
+		g.Expect(err).To(BeNil())
+
+		viper.SetConfigFile(cfg.Name())
+		q.Prefs = viper
 
 		g.Expect(q.Run(nil, []string{"why", "//cmd/aspect/query:query", "@com_github_bazelbuild_bazelisk//core:go_default_library"})).Should(Succeed())
 	})
@@ -60,7 +69,7 @@ func TestQuery(t *testing.T) {
 		spawner := bazel_mock.NewMockBazel(ctrl)
 		spawner.
 			EXPECT().
-			Spawn([]string{"cquery", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
+			Spawn([]string{"query", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
 			Return(0, nil)
 
 		promptRunner := query_mock.NewMockPromptRunner(ctrl)
@@ -77,12 +86,29 @@ func TestQuery(t *testing.T) {
 				Times(1),
 		)
 
-		q := &cquery.CQuery{
+		confirmationRunner := query_mock.NewMockConfirmationRunner(ctrl)
+		gomock.InOrder(
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("Y", nil).
+				Times(1),
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("Y", nil).
+				Times(1),
+		)
+
+		q := &query.Query{
 			Streams:       streams,
 			Bzl:           spawner,
 			IsInteractive: true,
 			Prompt: func(label string) shared.PromptRunner {
 				return promptRunner
+			},
+			Confirmation: func(question string) shared.ConfirmationRunner {
+				return confirmationRunner
 			},
 		}
 		q.Presets = []*shared.PresetQuery{
@@ -90,10 +116,17 @@ func TestQuery(t *testing.T) {
 				Name:        "why",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?target, ?dependency)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 		}
-		err := q.Run(nil, []string{"why"})
+
+		viper := *viper.New()
+		cfg, err := os.CreateTemp(os.Getenv("TEST_TMPDIR"), "cfg***.ini")
+		g.Expect(err).To(BeNil())
+
+		viper.SetConfigFile(cfg.Name())
+		q.Prefs = viper
+		err = q.Run(nil, []string{"why"})
 		g.Expect(err).To(BeNil())
 	})
 
@@ -125,12 +158,29 @@ func TestQuery(t *testing.T) {
 				Times(1),
 		)
 
-		q := &cquery.CQuery{
+		confirmationRunner := query_mock.NewMockConfirmationRunner(ctrl)
+		gomock.InOrder(
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("Y", nil).
+				Times(1),
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("Y", nil).
+				Times(1),
+		)
+
+		q := &query.Query{
 			Streams:       streams,
 			Bzl:           spawner,
 			IsInteractive: true,
 			Prompt: func(label string) shared.PromptRunner {
 				return promptRunner
+			},
+			Confirmation: func(question string) shared.ConfirmationRunner {
+				return confirmationRunner
 			},
 		}
 		q.Presets = []*shared.PresetQuery{
@@ -138,10 +188,18 @@ func TestQuery(t *testing.T) {
 				Name:        "why",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?target, ?dependency)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 		}
-		err := q.Run(cmd, []string{"why"})
+
+		viper := *viper.New()
+		cfg, err := os.CreateTemp(os.Getenv("TEST_TMPDIR"), "cfg***.ini")
+		g.Expect(err).To(BeNil())
+
+		viper.SetConfigFile(cfg.Name())
+		q.Prefs = viper
+
+		err = q.Run(cmd, []string{"why"})
 		g.Expect(err).To(MatchError(expectedError))
 	})
 
@@ -156,7 +214,7 @@ func TestQuery(t *testing.T) {
 		spawner := bazel_mock.NewMockBazel(ctrl)
 		spawner.
 			EXPECT().
-			Spawn([]string{"cquery", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
+			Spawn([]string{"query", "somepath(//cmd/aspect/query:query, @com_github_bazelbuild_bazelisk//core:go_default_library)"}).
 			Return(0, nil)
 
 		promptRunner := query_mock.NewMockPromptRunner(ctrl)
@@ -173,6 +231,20 @@ func TestQuery(t *testing.T) {
 				Times(1),
 		)
 
+		confirmationRunner := query_mock.NewMockConfirmationRunner(ctrl)
+		gomock.InOrder(
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("N", fmt.Errorf("")).
+				Times(1),
+			confirmationRunner.
+				EXPECT().
+				Run().
+				Return("N", fmt.Errorf("")).
+				Times(1),
+		)
+
 		selectRunner := query_mock.NewMockSelectRunner(ctrl)
 		selectRunner.
 			EXPECT().
@@ -180,16 +252,20 @@ func TestQuery(t *testing.T) {
 			Return(1, "", nil).
 			Times(1)
 
-		q := &cquery.CQuery{
+		q := &query.Query{
 			Streams:       streams,
 			Bzl:           spawner,
 			IsInteractive: true,
 			Prompt: func(label string) shared.PromptRunner {
+				fmt.Println("label:", label)
 				g.Expect(strings.Contains(label, "targettwo") || strings.Contains(label, "dependencytwo")).To(Equal(true))
 				return promptRunner
 			},
 			Select: func(presetNames []string) shared.SelectRunner {
 				return selectRunner
+			},
+			Confirmation: func(question string) shared.ConfirmationRunner {
+				return confirmationRunner
 			},
 		}
 		q.Presets = []*shared.PresetQuery{
@@ -197,22 +273,29 @@ func TestQuery(t *testing.T) {
 				Name:        "why1",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?targetone, ?dependencyone)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 			{
 				Name:        "why2",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?targettwo, ?dependencytwo)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 			{
 				Name:        "why3",
 				Description: "Determine why a target depends on another",
 				Query:       "somepath(?targetthree, ?dependencythree)",
-				Verb:        "cquery",
+				Verb:        "query",
 			},
 		}
-		err := q.Run(nil, []string{})
+
+		viper := *viper.New()
+		cfg, err := os.CreateTemp(os.Getenv("TEST_TMPDIR"), "cfg***.ini")
+		g.Expect(err).To(BeNil())
+
+		viper.SetConfigFile(cfg.Name())
+		q.Prefs = viper
+		err = q.Run(nil, []string{})
 		g.Expect(err).To(BeNil())
 	})
 }
