+++ pkg/aspect/query/shared/BUILD.bazel
@@ -1,14 +1,15 @@
 load("@io_bazel_rules_go//go:def.bzl", "go_library")
 
 go_library(
-    name = "cquery",
-    srcs = ["cquery.go"],
-    importpath = "aspect.build/cli/cmd/aspect/cquery",
+    name = "shared",
+    srcs = ["query.go"],
+    importpath = "aspect.build/cli/pkg/aspect/query/shared",
     visibility = ["//visibility:public"],
     deps = [
-        "//pkg/aspect/cquery",
+        "//pkg/aspecterrors",
         "//pkg/bazel",
         "//pkg/ioutils",
+        "@com_github_manifoldco_promptui//:promptui",
         "@com_github_spf13_cobra//:cobra",
     ],
 )
