+++ pkg/aspect/query/shared/mock/doc.go
